Move internal/autify.go to pkg/client/main.go
