+++ pkg/client/main.go
@@ -1,4 +1,4 @@
-package internal
+package client
 
 import (
 	"encoding/json"
